Raise parse errors for invalid `auth-source-xoauth2-creds' files.
Previously the code was calling `error' incorrectly, apparently a
typo?  The first arg is a format string, but the code was passing
it the result of calling `message' (which has no documented return
value).  It also passed a second nil arg for some reason.  ;-)

The result was that no error was raised, so it took me a while to
figure out that I had a typo in my creds file.